Move sanitization logic to Go served by WASM
Closes #19.

This change ensures that the sanitization is done in the front end browser
and also ensures that we can re-use the same code else where in the future, Eg: CLI tool, library etc.
It also opens up the possibility of using libraries/features available for the backend languages in the
frontend/browser.