Fix issues when no file input is chosen or an already sanitized file is chosen.
Closes #12.

When no file is chosen and the file input dialog is closed, the Go WASM code errors out and exits.
This prevents any other file from being loaded and sanitized. Work around is to refresh the page so that the Go WASM code loads again.
The desired behavior is, even if no file is selected, the Go WASM code doesn't errr out/exit and we're able to continue to load new
files to sanitize without refreshing the page.

When an already sanitized file is opened, it sanitizes it with new secret replacements.
The desired behavior is to detect if a field has already been sanitized and skip/sanitize accordingly.