Parallelize the secret detection logic
Closes #25.

Parallelize the secret detection logic by running each detection rule as
a separate go routine and aggregating the results in a channel.
The actual sanitization cannot be parallelized as it needs to be done sequentially
to ensure rule precedence i.e. the rule that is specified last overwrites the behavior of any matches from previous rules.