Support deterministic contextual replacement process
Closes #44.
Due to concurrent contextual replacement of different rules,
we cannot guarantee the order of execution and the output
cannot be deterministically singular. This makes it hard to test.
This change ensures that the contextual replacement is deterministic
by replacing the numeric siffix with a SHA-256 hash of the actual content
being sanitized. SHA-256 is still a NIST approved hash algorithm and the
chances for collision is negligible, so we need not worry about it at the moment.
https://csrc.nist.gov/projects/hash-functions#:~:text=algorithms%3A%20SHA%2D224%2C-,SHA%2D256,-%2C%20SHA%2D384%2C%20SHA